# 24x faster Bun.hash.crc32?