--- a/src/deps/zlib.posix.zig
+++ b/src/deps/zlib.posix.zig
@@ -87,3 +87,6 @@ pub inline fn inflateInit2(strm: anytype, windowBits: anytype) ReturnCode {
 pub inline fn inflateBackInit(strm: anytype, windowBits: anytype, window: anytype) ReturnCode {
     return inflateBackInit_(strm, windowBits, window, zlibVersion(), @import("std").zig.c_translation.cast(c_int, @import("std").zig.c_translation.sizeof(z_stream)));
 }
+
+pub extern fn adler32(uLong, ptr: ?[*]const u8, len: uInt) uLong;
+pub extern fn crc32(uLong, ptr: ?[*]const u8, len: uInt) uLong;